wallet, rpc: avoid descriptor range overflows
Seed commit: multiple historical descriptor-range hardening commits; current audit found remaining int32 boundary assumptions in RPC and descriptor wallet range handling.

Current paths: EvalDescriptorStringOrObject() expands RPC scan ranges through int positions; ProcessDescriptorImport() stores inclusive RPC ranges as exclusive wallet range_end; DescriptorScriptPubKeyMan::TopUpWithDB() adds next_index and target keypool size into int32 range_end; UpdateWalletDescriptor() reset m_max_cached_index to -1 even for descriptors whose range_start was already positive.

Reproducer: request a descriptor range ending at INT32_MAX through scantxoutset and importdescriptors, and use active ranged descriptors near the int32 boundary. Before the fix, inclusive-to-exclusive conversion and keypool top-up arithmetic could overflow int32 or cause cache bookkeeping to believe indexes below range_start were cached.

Fix: iterate RPC scan expansion with int64 while casting only after ParseDescriptorRange bounds, reject import ranges that cannot be represented as exclusive int32 range_end, check keypool top-up addition before narrowing, and initialize/reset m_max_cached_index to range_start - 1 for nonzero starts.

Tests: rpc_scantxoutset.py covers the allowed single INT32_MAX scan expansion; wallet_importdescriptors.py checks importdescriptors rejects INT32_MAX as an inclusive range end because wallet storage needs end + 1.

Validation: cmake --build build_audit_wallet --target bitcoind -j2; test/functional/rpc_scantxoutset.py --configfile=build_audit_wallet/test/config.ini; test/functional/wallet_importdescriptors.py --configfile=build_audit_wallet/test/config.ini; git diff --check.

+++ src/rpc/util.cpp
@@ -1364,14 +1364,14 @@ std::vector<CScript> EvalDescriptorStringOrObject(const UniValue& scanobject, Fl
         range.second = 0;
     }
     std::vector<CScript> ret;
-    for (int i = range.first; i <= range.second; ++i) {
+    for (int64_t i = range.first; i <= range.second; ++i) {
         for (const auto& desc : descs) {
             std::vector<CScript> scripts;
-            if (!desc->Expand(i, provider, scripts, provider)) {
+            if (!desc->Expand(static_cast<int>(i), provider, scripts, provider)) {
                 throw JSONRPCError(RPC_INVALID_ADDRESS_OR_KEY, strprintf("Cannot derive script without private keys: '%s'", desc_str));
             }
             if (expand_priv) {
-                desc->ExpandPrivate(/*pos=*/i, provider, /*out=*/provider);
+                desc->ExpandPrivate(/*pos=*/static_cast<int>(i), provider, /*out=*/provider);
             }
             std::move(scripts.begin(), scripts.end(), std::back_inserter(ret));
         }
